backend/db/requirements.txt
python-dotenv
plaid-python
requests
Flask
Flask-SQLAlchemy